Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv

# ML / AI
scikit-learn
xgboost
shap
pandas
numpy
transformers
torch
spacy
supabase
python-dotenv
google-generativeai
requests
faiss-cpu
sentence-transformers
